readjust permissions and readjust sidebar

+++ resources/views/employeeDocumentTypes/index.blade.php
@@ -51,7 +51,7 @@ $profile=\App\Models\Utility::get_file('uploads/avatar');
                             <thead>
                             <tr>
                                 <th>{{__('Name')}}</th>
-                                @if(Gate::check('edit document type') || Gate::check('delete document type'))
+                                @if(Gate::check('edit document') || Gate::check('delete document'))
                                     <th>{{__('Action')}}</th>
                                 @endif
                             </tr>
@@ -62,19 +62,19 @@ $profile=\App\Models\Utility::get_file('uploads/avatar');
                                     <td>{{ $document_type->name }}</td>
                                     {{-- @if(Gate::check('edit document type') || Gate::check('delete document type')) --}}
                                         <td>
-                                            {{-- @can('edit document type') --}}
+                                            @can('edit document')
                                                 <div class="action-btn bg-primary ms-2">
                                                     <a href="#" data-url="{{ route('employee-document-types.edit',$document_type->id)}}" data-size="lg" data-ajax-popup="true" data-title="{{__('Edit Document Type')}}" class="mx-3 btn btn-sm align-items-center" data-bs-toggle="tooltip" title="{{__('Edit')}}" data-original-title="{{__('Edit')}}"><i class="ti ti-pencil text-white"></i></a>
                                                 </div>
-                                            {{-- @endcan --}}
-                                            {{-- @can('delete document type') --}}
+                                            @endcan
+                                            @can('delete document')
                                                 <div class="action-btn bg-danger ms-2">
                                                     {!! Form::open(['method' => 'DELETE', 'route' => ['employee-document-types.destroy', $document_type->id],'id'=>'delete-form-'.$document_type->id]) !!}
 
                                                     <a href="#" class="mx-3 btn btn-sm align-items-center bs-pass-para" data-bs-toggle="tooltip" title="{{__('Delete')}}" data-original-title="{{__('Delete')}}" data-confirm="{{__('Are You Sure?').'|'.__('This action can not be undone. Do you want to continue?')}}" data-confirm-yes="document.getElementById('delete-form-{{$document_type->id}}').submit();"><i class="ti ti-trash text-white"></i></a>
                                                     {!! Form::close() !!}
                                                 </div>
-                                            {{-- @endif --}}
+                                            @endif
                                         </td>
                                     {{-- @endif --}}
                                 </tr>
