+++ resources/views/shipment/all_shipment.blade.php
@@ -9,11 +9,12 @@
 
 @section('action-btn')
     <div class="float-end">
-       
-        <a href="{{ route('shipment.create') }}"
-            class="btn btn-sm btn-primary">
-            <i class="ti ti-plus"></i>
-        </a>
+       @can("create shipment")
+       <a href="{{ route('shipment.create') }}"
+           class="btn btn-sm btn-primary">
+           <i class="ti ti-plus"></i>
+       </a>
+       @endcan
     </div>
 @endsection
 
@@ -125,19 +126,20 @@
                                     <td class="font-style">{{ $shipment->package_status }}</td>
                                     
                                     <td>
+                                        @can("edit shipment")
+                                            <div class="action-btn bg-primary ms-2">
+                                                <a href="{{route('shipment.edit',\Illuminate\Support\Facades\Crypt::encrypt($shipment->id))}}" class="mx-3 btn btn-sm align-items-center" data-bs-toggle="tooltip" title="{{__('Edit')}}"
+                                                data-original-title="{{__('Edit')}}"><i class="ti ti-pencil text-white"></i></a>
+                                            </div>    
+                                        @endcan
+                                        @can("delete shipment")
+                                            <div class="action-btn bg-danger ms-2">
+                                            {!! Form::open(['method' => 'DELETE', 'route' => ['shipment.destroy', $shipment->id],'id'=>'delete-form-'.$shipment->id]) !!}
 
-                                                <div class="action-btn bg-primary ms-2">
-                                                    <a href="{{route('shipment.edit',\Illuminate\Support\Facades\Crypt::encrypt($shipment->id))}}" class="mx-3 btn btn-sm align-items-center" data-bs-toggle="tooltip" title="{{__('Edit')}}"
-                                                    data-original-title="{{__('Edit')}}"><i class="ti ti-pencil text-white"></i></a>
-                                                </div>
-
-                                                    
-                                                <div class="action-btn bg-danger ms-2">
-                                                {!! Form::open(['method' => 'DELETE', 'route' => ['shipment.destroy', $shipment->id],'id'=>'delete-form-'.$shipment->id]) !!}
-
-                                                    <a href="#" class="mx-3 btn btn-sm align-items-center bs-pass-para" data-bs-toggle="tooltip" title="{{__('Delete')}}" data-original-title="{{__('Delete')}}" data-confirm="{{__('Are You Sure?').'|'.__('This action can not be undone. Do you want to continue?')}}" data-confirm-yes="document.getElementById('delete-form-{{$shipment->id}}').submit();"><i class="ti ti-trash text-white"></i></a>
-                                                    {!! Form::close() !!}
-                                                </div>
+                                                <a href="#" class="mx-3 btn btn-sm align-items-center bs-pass-para" data-bs-toggle="tooltip" title="{{__('Delete')}}" data-original-title="{{__('Delete')}}" data-confirm="{{__('Are You Sure?').'|'.__('This action can not be undone. Do you want to continue?')}}" data-confirm-yes="document.getElementById('delete-form-{{$shipment->id}}').submit();"><i class="ti ti-trash text-white"></i></a>
+                                                {!! Form::close() !!}
+                                            </div>
+                                        @endcan
 
                                         </td>
                                 </tr>
